Rename GetCalendarElement to GetStyledCalendarElement
