+++ src/utils/CalendarUtils/CalendarUtils.js
@@ -1,5 +1,5 @@
 import { stringify } from 'svgson';
-import * as GetCalendarElement from '../GetCalendarElement/GetCalendarElement';
+import * as GetStyledCalendarElement from '../GetStyledCalendarElement/GetStyledCalendarElement';
 import * as JavaScriptUtils from '../JavaScriptUtils/JavaScriptUtils';
 
 export const RequiredParamsExist = (container, gitHubUsers) => {
@@ -15,8 +15,8 @@ export const RequiredParamsExist = (container, gitHubUsers) => {
 };
 
 export const RenderCalendarWithContributions = (container, calendar, totalContributions) => {
-  const calendarContainer = GetCalendarElement.container(container);
-  const calendarHeader = GetCalendarElement.header(totalContributions);
+  const calendarContainer = GetStyledCalendarElement.container(container);
+  const calendarHeader = GetStyledCalendarElement.header(totalContributions);
 
   const stringifiedHTMLContent = stringify(calendar);
