+++ src/utils/GetStyledCalendarElement/GetStyledCalendarElement.test.js
@@ -2,13 +2,13 @@ import { expect } from 'chai';
 import jsdom from 'mocha-jsdom';
 import proxyquire from 'proxyquire';
 
-const GetCalendarElement = proxyquire('./GetCalendarElement.js', {
+const GetStyledCalendarElement = proxyquire('./GetStyledCalendarElement.js', {
   elly: () => ({
     style: {},
   }),
 });
 
-describe('GetCalendarElement', () => {
+describe('GetStyledCalendarElement', () => {
   // https://github.com/rstacruz/mocha-jsdom/issues/36
   // https://github.com/jsdom/jsdom/issues/2383
   jsdom({
@@ -17,7 +17,7 @@ describe('GetCalendarElement', () => {
 
   describe('container', () => {
     it('returns a 700px width element', () => {
-      const calendarContainer = GetCalendarElement.container('.class');
+      const calendarContainer = GetStyledCalendarElement.container('.class');
 
       expect(calendarContainer.style.width).to.equal('700px');
     });
@@ -25,13 +25,13 @@ describe('GetCalendarElement', () => {
 
   describe('header', () => {
     it('returns a `div` element', () => {
-      const calendarHeader = GetCalendarElement.header();
+      const calendarHeader = GetStyledCalendarElement.header();
 
       expect(calendarHeader.nodeName).to.equal('DIV');
     });
 
     it('appends two child nodes to the header', () => {
-      const calendarHeader = GetCalendarElement.header();
+      const calendarHeader = GetStyledCalendarElement.header();
 
       expect(calendarHeader.childNodes.length).to.equal(2);
     });
